+++ modules/tx_unpacker/Cargo.toml
@@ -20,7 +20,6 @@ futures = "0.3.31"
 hex = { workspace = true }
 pallas = { workspace = true }
 rayon = { workspace = true }
-thiserror = "2.0.17"
 tracing = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
