+++ modules/utxo_state/src/utxo_state.rs
@@ -49,7 +49,7 @@ use fake_immutable_utxo_store::FakeImmutableUTXOStore;
 
 use crate::reference_scripts_state::ReferenceScriptsState;
 mod utils;
-mod validations;
+pub mod validations;
 
 declare_cardano_reader!(
     GenesisCompleteReader,
