+++ modules/utxo_state/src/validations/mod.rs
@@ -12,6 +12,7 @@ use anyhow::Result;
 use crate::utils;
 mod alonzo;
 mod babbage;
+pub mod phase2;
 mod phase_two;
 mod shelley;
