+++ modules/utxo_state/src/validations/phase2.rs
@@ -9,18 +9,11 @@
 //! It verifies that all scripts in a transaction execute successfully within their
 //! allocated execution budgets.
 //!
-//! # Feature Flag
-//!
-//! Phase 2 validation is disabled by default. Enable it via configuration:
-//! ```toml
-//! [module.tx-unpacker]
-//! phase2_enabled = true
-//! ```
 //!
 //! # Example
 //!
 //! ```ignore
-//! use acropolis_module_tx_unpacker::validations::phase2::{
+//! use acropolis_module_utxo_state::validations::phase2::{
 //!     evaluate_script, ExUnits, PlutusVersion,
 //! };
 //!
@@ -904,6 +897,35 @@ pub fn script_lang_to_plutus_version(script_lang: &ScriptLang) -> Option<PlutusV
 /// Convert from Phase2Error to common::Phase2ValidationError for integration.
 impl From<Phase2Error> for acropolis_common::validation::Phase2ValidationError {
     fn from(err: Phase2Error) -> Self {
-        Self::UplcMachineError(err.to_string())
+        acropolis_common::validation::Phase2ValidationError::UplcMachineError(match err {
+            Phase2Error::ScriptFailed(script_hash, message) => {
+                acropolis_common::validation::UplcMachineError::ScriptFailed {
+                    script_hash,
+                    message,
+                }
+            }
+            Phase2Error::BudgetExceeded(script_hash, cpu, mem) => {
+                acropolis_common::validation::UplcMachineError::BudgetExceeded {
+                    script_hash,
+                    cpu,
+                    mem,
+                }
+            }
+            Phase2Error::DecodeFailed(script_hash, reason) => {
+                acropolis_common::validation::UplcMachineError::DecodeFailed {
+                    script_hash,
+                    reason,
+                }
+            }
+            Phase2Error::MissingScript(index) => {
+                acropolis_common::validation::UplcMachineError::MissingScript { index }
+            }
+            Phase2Error::MissingDatum(datum_hash) => {
+                acropolis_common::validation::UplcMachineError::MissingDatum { datum_hash }
+            }
+            Phase2Error::MissingRedeemer(script_hash) => {
+                acropolis_common::validation::UplcMachineError::MissingRedeemer { script_hash }
+            }
+        })
     }
 }
