+++ modules/utxo_state/tests/calibration.rs
@@ -6,7 +6,7 @@
 
 mod common;
 
-use acropolis_module_tx_unpacker::validations::phase2::evaluate_raw_flat_program;
+use acropolis_module_utxo_state::validations::phase2::evaluate_raw_flat_program;
 use common::{
     calibrate, get_calibration, CalibrationBaseline, CALIBRATION_ITERATIONS, CALIBRATION_SCRIPT,
 };
