+++ processes/omnibus/omnibus-preview.toml
@@ -5,9 +5,9 @@
 # ============================================================================
 [global.startup]
 network-name = "preview"
-startup-mode = "genesis"   # Options: "genesis" | "snapshot"
-sync-mode = "mithril"      # Options: "mithril" | "upstream"
-block-flow-mode = "direct" # Options: "direct" | "consensus"
+startup-mode = "genesis"      # Options: "genesis" | "snapshot"
+sync-mode = "mithril"         # Options: "mithril" | "upstream"
+block-flow-mode = "consensus" # Options: "direct"  | "consensus"
 topic = "cardano.sequence.start"
 
 # ============================================================================
@@ -73,14 +73,6 @@ force-validation = false
 # Subscriptions needed for validation
 bootstrapped-subscribe-topic = "cardano.sequence.bootstrapped"
 protocol-parameters-subscribe-topic = "cardano.protocol.parameters"
-
-# Plutus Phase 2 Validation
-# Enable Phase 2 script validation (Plutus/native scripts).
-# When enabled, transactions with Plutus scripts will be fully evaluated.
-# When disabled (default), script validation is skipped for faster sync.
-# WARNING: Enable only after thorough testing. Incorrect script evaluation
-# can cause consensus failures or missed invalid transactions.
-phase2-enabled = false
 
 publish-utxo-deltas-topic = "cardano.utxo.deltas"
 publish-asset-deltas-topic = "cardano.asset.deltas"
